third_party/WebKit/LayoutTests/fast/dom/Element/class-list-expected.txt

Tests that classList works on non HTMLElements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS htmlElement.classList instanceof DOMTokenList is true
PASS svgElement.classList instanceof DOMTokenList is true
PASS xmlElement.classList instanceof DOMTokenList is true
PASS successfullyParsed is true

TEST COMPLETE
